+++ test/gurka/test_matrix.cc
@@ -766,11 +766,41 @@ TEST(StandAlone, MatrixSecondPass) {
     EXPECT_FALSE(api.matrix().second_pass(1));
     EXPECT_GT(api.matrix().times(2), 0.f);
     EXPECT_TRUE(api.matrix().second_pass(2));
+    EXPECT_GT(api.matrix().distances(2), api.matrix().distances(1));
     EXPECT_GT(api.matrix().times(2), api.matrix().times(1));
 
     // I -> I & K -> K shouldn't be processed a second time either
     EXPECT_FALSE(api.matrix().second_pass(0));
     EXPECT_FALSE(api.matrix().second_pass(3));
     EXPECT_TRUE(api.info().warnings(0).description().find('2') != std::string::npos);
+  }
+}
+
+TEST(StandAlone, CostMatrixTrivialRoutes) {
+  const std::string ascii_map = R"(
+    A---B--2->-1--C---D
+        |         |
+        |         |
+        E--3---4--F
+  )";
+  const gurka::ways ways = {
+      {"AB", {{"highway", "residential"}}}, {"BC", {{"highway", "residential"}, {"oneway", "yes"}}},
+      {"CD", {{"highway", "residential"}}}, {"BE", {{"highway", "residential"}}},
+      {"EF", {{"highway", "residential"}}}, {"FC", {{"highway", "residential"}}},
+  };
+  const auto layout = gurka::detail::map_to_coordinates(ascii_map, 100);
+  auto map =
+      gurka::buildtiles(layout, ways, {}, {}, VALHALLA_BUILD_DIR "test/data/costmatrix_trivial");
+
+  // test the against-oneway case
+  {
+    auto matrix = gurka::do_action(valhalla::Options::sources_to_targets, map, {"1"}, {"2"}, "auto");
+    EXPECT_EQ(matrix.matrix().distances(0), 2200);
+  }
+
+  // test the normal trivial case
+  {
+    auto matrix = gurka::do_action(valhalla::Options::sources_to_targets, map, {"3"}, {"4"}, "auto");
+    EXPECT_EQ(matrix.matrix().distances(0), 400);
   }
 }